Fix filepath encoding issue and split headers
