+++ src_for_visualstudio/DatabaseSystemProject/DatabaseSystemProject/DatabaseSystemProject.vcxproj
@@ -143,6 +143,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="MapReduce.h" />
+    <ClInclude Include="WordCount.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
