+++ src_for_visualstudio/DatabaseSystemProject/DatabaseSystemProject/MapReduce.h
@@ -10,6 +10,28 @@
 #include <cctype>
 #include <queue>
 
+template <typename T>
+class ConcurrentQueue {
+private:
+    std::queue<T> queue;
+    std::mutex mtx;
+public:
+    void push(T element) {
+        std::lock_guard<std::mutex> lock(mtx);
+        queue.push(std::move(element));
+    }
+
+    bool try_pop(T& popped_value) {
+        std::lock_guard<std::mutex> lock(mtx);
+        if (queue.empty()) {
+            return false;
+        }
+        popped_value = queue.front();
+        queue.pop();
+        return true;
+    }
+};
+
 class IMapper {
 public:
     virtual ~IMapper() {}
